pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dropbox-cleaner"
version = "1.0.0"
description = "Local-first Dropbox inventory and staged archive desktop utility"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
license-files = ["LICENSE"]
authors = [
  { name = "Dropbox Cleaner contributors" }
]
keywords = ["dropbox", "desktop", "archive", "inventory", "tkinter", "cli"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Environment :: Win32 (MS Windows)",
  "Intended Audience :: End Users/Desktop",
  "Intended Audience :: Information Technology",
  "Operating System :: MacOS",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Topic :: Desktop Environment",
  "Topic :: System :: Archiving",
  "Topic :: Utilities",
]
dependencies = [
  "dropbox>=12.0.2,<13",
  "keyring>=25.0.0,<26",
  "platformdirs>=4.0.0,<5",
  "PyYAML>=6.0.0,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<10",
  "build>=1.2.0,<2",
]

[project.scripts]
dropbox-cleaner = "app.cli.main:main"
dropbox-cleaner-gui = "app.ui.main:main"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
